Mods for unsigned long long (EvtEventId)

+++ GlastClassify/src/AtwoodTrees.cxx
@@ -202,8 +202,10 @@ bool AtwoodTrees::execute()
     double calenergy    = *m_CalEnergyRaw;
     double calCsiRln    = *m_CalCsIRLn;
 
-    int    eventId      = *m_eventId;
-    int    run          = *m_run;
+// LSR 14-Jul-08 code for ntuple types; potential changes here!
+
+    unsigned long long eventId = *m_eventId;
+    int                run     = *m_run;
 
     // Always zero the CTB output values in case cuts below fail
     m_treeAnalysis->zeroCTvals();
